feat(steam-track,sim-ring): add Perfetto tracker MVP
Support for creating the required Perfetto Protobuf messages is provided
via the new perfetto_trace_builder module within steam-track. This
module enables code reuse with the tracker and future use cases such as
being able to post-process steam-track data into Perfetto traces.

The Perfetto tracker is considered to be the MVP, as whilst it currently
creates traces well suited for understanding data such as buffer
fullness from a simulation, it only creates Perfetto TrackEvents for
steam-track's Track enter and exit events.

The sim-ring example application has been extended to support the use of
the Perfetto tracker, allowing it to generate Perfetto traces, as well
as the previously supported text and binary trace formats supported by
steam-spotter.

This new functionality of steam-track is guarded by the `perfetto`
feature. With this new feature enabled the build of steam-track will
cause the steam-perfetto-sys crate build to download Perfetto from
source, and the steam-perfetto crate to generate the required Protobuf
accessor functions.

As the `_perfetto-protoc` feature of steam-perfetto is not enabled by
steam-track a Protobuf compiler must already available for use on the
path.