docs: detail the use of cargo-about
The generation of the licenses.html file when the pre-commit hooks run
cargo-about is now described in the "Committing a Change" section which
appears in both the README.md and the "Setup" chapter of the developer
guide.

The generated licenses.html file is now included in the built developer
guide output, and the link in README.md will direct readers to the
copy included in the Github Pages hosted version of the guide.