infra: perform deterministic `cargo install` builds
Performing `--locked` Cargo builds ensures that tools will only be
rebuilt when new releases are made, rather than also when new
sematically compatible releases of their dependencies are avaliable.

The install of Cocogitto is now limited to just the `cog` binary used.
Previously the whole of Cocogitto was built everytime due to the
following Cargo bug:
https://github.com/rust-lang/cargo/issues/8703

The combination of these two changes should mean that often there is
nothing to rebuild having restored the cached directories.