feat(steam-track): simplify to remove concept of trace enable
We had two concepts for steam-track; the log level and whether trace was enabled. However,
this distinction was unhelpful so now we just have the level and tracing is enabled whenever
logging at level log::Level::Trace is set.

Also, in order to support different behaviour with a MultiTracker we needed to move the
level/enable caching to the EntityManager rather than each entity.