test(steam-config): remove use of escargot library
Several instances of these tests failing with the error `"No such file
or directory"` have been observed during CI, with the tests passing when
the workflow is re-run.

Instead of using escargot the tests now run the required cargo commands
directly, as hopefully this issue is to do with the way escargot calls
for required binaries to be built (as it claims to offer reduced compile
times) and not an issue with the way Cargo locks the build directory.