test(steam-config): serialise tests which share an executable
Several instances of these tests failing with the error `"No such file
or directory"` have been observed during CI, with the tests passing when
the workflow is re-run. The removal of escargot did not resolved this.

I suspect the race condition may be caused by multiple tests attempting
to build the same executable:
- Cargo will lock the target directory preventing multiple compilations
  from occuring in parallel.
- The build directory is unlocked once the binary has been built.
- Another test manages to lock the target directory and rebuild the
  binary whilst the first `cargo run` command attempts to launch process
  running the binary it built.

However it is not clear why subsequent `cargo run` invocations should
build the binary as the source cannot have changed.