fix: address all clippy::cargo lints
By default the lints in the clippy::cargo group are set to "allow" and
as such do not emit any warnings or errors, but can be set to "deny" by
running `cargo clippy -- -D clippy::cargo`.

The lint that checks for multiple crate versions within the dependency
tree is explicitly set to "allow".

The `cargo clippy-strict` alias has been updated to include this group
as warnings, which will be promoted to errors when run via the
pre-commit hook.