fix: address some of the clippy::pedantic lints
By default the lints in the clippy::pedantic group are set to "allow"
and as such do not emit any warnings or errors, but can be set to "warn"
by running `cargo clippy -- -W clippy::pedantic`.

The clippy::pedantic lint group contains many lints, not all of which
have been addresses yet. The following individual lints have been fixed:

- clippy::explicit-iter-loop
- clippy::manual-string-new
- clippy::must-use-candidate
- clippy::needless-pass-by-value
- clippy::semicolon-if-nothing-returned
- clippy::single-char-pattern
- clippy::uninlined-format-args
- clippy::unnecessary-semicolon
- clippy::unused-self

The following individual lints are still to be addressed (either fixed
or explicitly allowed):

- clippy::assigning-clones
- clippy::cast-lossless
- clippy::cast-possible-truncation
- clippy::cast-precision-loss
- clippy::cast-sign-loss
- clippy::doc-markdown
- clippy::expl-impl-clone-on-copy
- clippy::explicit-into-iter-loop
- clippy::float-cmp
- clippy::ignored-unit-patterns
- clippy::implicit-clone
- clippy::inconsistent-struct-constructor
- clippy::manual-assert
- clippy::manual-let-else
- clippy::match-same-arms
- clippy::match-wild-err-arm
- clippy::missing-errors-doc
- clippy::missing-fields-in-debug
- clippy::missing-panics-doc
- clippy::ptr-as-ptr
- clippy::range-plus-one
- clippy::redundant-closure-for-method-calls
- clippy::return-self-not-must-use
- clippy::similar-names
- clippy::single-match-else
- clippy::struct-excessive-bools
- clippy::struct-field-names
- clippy::trivially-copy-pass-by-ref
- clippy::unnecessary-wraps
- clippy::unnested-or-patterns
- clippy::unused-async

The `cargo clippy-strict` alias has been updated to include the
clippy::pedantic group as warnings, which will be promoted to errors
when run via the pre-commit hook. The remaining lints that are not yet
addresses are explicitly set to "allow".