docs: ensure `cargo doc --document-private-items` can be used
As it can be useful when developing functionality withing the STEAM
packages to build the API documentation with `--document-private-items`
enabled all warnings and errors reported with this set have been
resolved.

A new pre-commit hook has been added to ensure that documentation builds
of this kind remain warning and error free.

Other minor documentation improvements are also included, such as
spelling corrections.