infra: update stable Rust toolchain to 1.88.0
Updating the toolchain required a fix to steam-track as the
`uninlined_format_args` lint was moved from the `pedantic` group to
the `style` group of lints, see
https://github.com/rust-lang/rust-clippy/pull/14160.

As the `clippy::all` group does not include the pedantic group, warnings
were emitted for the generated Cap'n Proto code. To disable all warnings
currently raised both clippy allow attributes are now required.