infra: lock build and development tool versions
Most tools are now fixed to specific versions to avoid unexpected CI
failures, either when opening a pull request or when a pull request is
merged, due to a new version of a tool having been released in the mean
time.

The versions of tools installed with `apt` and `brew` have not been
fixed.

The stable Rust toolchain used to build STEAM is now specified via the
rust-toolchain.toml file. This will cause `rustup` to automatically
activate the required toolchain when commands are run from within the
STEAM directory tree. `rustup` will automatically install the toolchain
as required when any command is invoked, hence executing `rustup show`
within the install-build-dependencies install script.

The rust-toolchain.toml file format does not support the controlling of
multiple toolchain installed. As STEAM also relies on a nightly Rust
toolchain to provide a version of `rustfmt` that allows the use of
"unstable" features the install-dev-dependencies install script still
installs this. To reduce the time taken the "minimal" profile of the
toolchain is now installed (plus `rustfmt` which is not included in
minimal).

The versions of tools installed with both `cargo` and `npm` are
specified directly in the `install.sh` scripts using the "@version"
notation they both support.