feat(steam-spotter): --perfetto option to generate MVP Perfetto trace
steam-spotter can now generate Perfetto trace output which can be
visualised using the web-based Perfetto UI. When generating a Perfetto
trace the usual TUI will not be started.

Support for creating the required Perfetto Protobuf messages is provided
via the perfetto_trace_builder module within steam-track.

The Perfetto generator is considered to be the MVP, as whilst it
currently creates traces well suited for understanding data such as
buffer fullness from a simulation, it only creates Perfetto TrackEvents
for steam-track's Track enter and exit events.

This new functionality of steam-track is guarded by the `perfetto`
feature, which is enabled by default. With this new feature enabled the
build of steam-track will cause the steam-perfetto-sys crate build to
download Perfetto from source, and the steam-perfetto crate to generate
the required Protobuf accessor functions.

As the `_perfetto-protoc` feature of steam-perfetto is not enabled by
steam-track a Protobuf compiler must already available for use on the
path.