feat(steam-perfetto-sys): allow source download during documentation-only build
The download of the Perfetto source enabled by the `_perfetto_src`
feature is a relatively quick operation, and provides the Protobuf
schemas dependent packages will most likely be relying on the
steam-perfetto-sys package for.

Documentation-only builds of packages that generate Rust source from
these schema will fail if the Perfetto source is not made avaliable as
expected when the feature is enabled.

This download process will now occur when the feature is enabled
regardless of whether or not a documentation-only build is occuring.