fix(steam-perfetto-sys): Avoid use of non-standard `-F` option
The FreeBSD (as used by macOS) and GNU implementations of `ln` both
support the `-F` option but unfortunately use it for completely
different behaviours.

When using the FreeBSD implementation `-F` implies `-f` (when `-i` is
not specified), and results the behaviour required here. The behaviour
of the `-f` option is common to both implementations, so is now set
explicitly instead.