chore: temporarily disable semver checks for steam-perfetto
The semantic version checks are performed using the main branch as the
base version as the packages are not currently published to a registry.

Unfortunately cargo-semver-checks cannot not handle the case of a new
package being added on a branch and therefore not being found on main.
To work around this limitation we must disable the checks for new
packages when they are first added.

This commit should be reverted once steam-perfetto has been merged onto
the main branch.