infra: set --locked when running cargo-about
Perviously `--offline` was used to avoid cargo-about attempting to query
https://clearlydefined.io/ as this was found to be unreliable and could
cause large variations in the number of license texts cargo-about
managed to fetch and include in licenses.html.

As of version 0.8.0 cargo-about no longer uses clearly-defined.io. See
https://github.com/EmbarkStudios/cargo-about/releases/tag/0.8.0.

Setting `--locked` instead ensures that cargo-about will use exactly the
same dependency versions as the build, but allow Cargo to download the
source of these as required to retrieve license texts as required.