infra: reduce macOS GitHub-hosted runner usage
The macOS runners are effectively charged at ~5x the Linux runners,
though very often appear to offer better runtime performance.

The check-semantic-versioning CI job is being switched to run on Linux
to reduce the risk of hitting usage limits again in future, which also
removes the need to run the `install-dev-dependencies` action on macOS
(and therefore also the `warm-dev-deps-cache` job).

The downside of this is that the install-dev-dependencies\install.sh
script will no longer be run on macOS by the CI system.