feat!: rebrand as TRAMWAY
Tramway is Rust Async Modelling for hardWare Architecture and sYstems.

The project has been rebranded to avoid name clashes when publishing
releases due to the existence of many packages named with the prefix
`steam-` already published on crates.io (as well as to avoid any
confusion with other well known pieces of software using the name).