feat!: rebrand as GWR
The Great Western Runtime.

A second attempt at rebranding the project, as the previous new name
(TRAMWAY) turned out be unergonomic in use, partly due to it's length.

The project has been rebranded to avoid name clashes when publishing
releases due to the existence of many packages named with the prefix
steam- already published on crates.io (as well as to avoid any
confusion with other well known pieces of software using the name).