feat: add routed fabric and improve spotter frontend visualisations
A fully routed fabric has been added using existing components and
no custom logic. It has the same external interfaces so it is a drop-in
replacement for the functional fabric.

The spotter frontend has changed to using the sunburst visualisation
as it is more robust to working with large designs.

The frontend also now calculates the radius of each element using the
area of its children. This results in better sizing of parent nodes.